Add a tryhard option to chain::generate().
A Boolean option, tryhard, has been added to both forms of
chain::generate().

When set to true, this option allows generate to continue producing
output when it stumbles upon the last prefix entry in the chain.  In
that event, a random prefix will be chosen to allow output generation
to continue.

When set to false, output generation will stop if the last prefix
comes up and we still want output.  This will lead to short output,
but was the behavior before the tryhard options was added.

The tryhard option has a default value of false.  You need not change
any code that previously used the markov library.  This new option
does change two function signatures in the library, so you will need
to recompile any libmarkov client code.

Signed-off-by: Jason Stephenson <jason@sigio.com>